feat: 부트캠프 정보 테이블 컴포넌트 분리
